% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_message}
\alias{append_message}
\title{Append message to a list}
\usage{
append_message(content, role = "user", x = NULL, ...)
}
\arguments{
\item{content}{The content of the message.}

\item{role}{The role of the message. Can be "user", "system", "assistant". Default is "user".}

\item{x}{A list of messages. Default is NULL.}

\item{...}{Additional arguments such as images.}
}
\value{
A list of messages with the new message appended.
}
\description{
Appends a message (add to end of a list) to a list of messages. The role and content will be converted to a list and appended to the input list.
}
\examples{
append_message("user", "Hello")
append_message("system", "Always respond nicely")
}
