% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_options.R
\name{check_option_valid}
\alias{check_option_valid}
\title{Check if an option is valid.}
\usage{
check_option_valid(opt)
}
\arguments{
\item{opt}{An option (character) to check.}
}
\value{
Returns TRUE if the option is valid, FALSE otherwise.
}
\description{
Check if an option is valid.
}
\examples{
check_option_valid("mirostat")
check_option_valid("invalid_option")
}
