% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_options.R
\name{check_options}
\alias{check_options}
\title{Check if a vector of options are valid}
\usage{
check_options(opts = NULL)
}
\arguments{
\item{opts}{A vector of options to check.}
}
\value{
Returns a list with two elements: valid_options and invalid_options.
}
\description{
Check if a vector of options are valid
}
\examples{
check_options(c("mirostat", "invalid_option"))
check_options(c("mirostat", "num_predict"))
}
