% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{create}
\alias{create}
\title{Create a model}
\usage{
create(
  model,
  from,
  system = NULL,
  stream = FALSE,
  endpoint = "/api/create",
  host = NULL
)
}
\arguments{
\item{model}{Name of the model to create.}

\item{from}{Name of an existing model to create the new model from.}

\item{system}{System prompt for the model. Default is NULL.}

\item{stream}{Enable response streaming. Default is FALSE.}

\item{endpoint}{The endpoint to create the model. Default is "/api/create".}

\item{host}{The base URL to use. Default is NULL, which uses Ollama's default base URL.}
}
\value{
A response in the format specified in the output parameter.
}
\description{
Create a model from another model, a safetensors directory (not implemented), or a GGUF file (not implemented).
}
\examples{
\dontshow{if (test_connection(logical = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
create("mario", "deepseek-r1:1.5b", system = "You are Mario from Super Mario Bros.")
model_avail("mario") # check mario model has been created
list_models() # mario model has been created
generate("mario", "who are you?", output = "text",
          timeout = 300)  # model should say it's Mario
delete("mario") # delete the model created above
model_avail("mario") # model no longer exists
\dontshow{\}) # examplesIf}
}
\references{
\href{https://github.com/ollama/ollama/blob/main/docs/api.md#create-a-model}{API documentation}
}
