% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_message}
\alias{create_message}
\title{Create a message}
\usage{
create_message(content, role = "user", ...)
}
\arguments{
\item{content}{The content of the message.}

\item{role}{The role of the message. Can be "user", "system", "assistant". Default is "user".}

\item{...}{Additional arguments such as images.}
}
\value{
A list of messages.
}
\description{
Create a message
}
\examples{
create_message("Hello", "user")
create_message("Always respond nicely", "system")
create_message("I am here to help", "assistant")
}
