% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_messages}
\alias{create_messages}
\title{Create a list of messages}
\usage{
create_messages(...)
}
\arguments{
\item{...}{A list of messages, each of list class.}
}
\value{
A list of messages, each of list class.
}
\description{
Create messages for \code{chat()} function.
}
\examples{
messages <- create_messages(
    create_message("be nice", "system"),
    create_message("tell me a 3-word joke")
)

messages <- create_messages(
    list(role = "system", content = "be nice"),
    list(role = "user", content = "tell me a 3-word joke")
)
}
