% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encode_images_in_messages}
\alias{encode_images_in_messages}
\title{Encode images in messages to base64 format}
\usage{
encode_images_in_messages(messages)
}
\arguments{
\item{messages}{A list of messages, each of list class. Generally used in the \code{chat()} function.}
}
\value{
A list of messages with images encoded in base64 format.
}
\description{
Encode images in messages to base64 format
}
\examples{
image <- file.path(system.file("extdata", package = "ollamar"), "image1.png")
messages <- list(
    list(role = "user", content = "what is in the image?", images = image)
)
messages_updated <- encode_images_in_messages(messages)
}
