% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{image_encode_base64}
\alias{image_encode_base64}
\title{Read image file and encode it to base64.}
\usage{
image_encode_base64(image_path)
}
\arguments{
\item{image_path}{The path to the image file.}
}
\value{
A base64 encoded string.
}
\description{
Read image file and encode it to base64.
}
\examples{
image_path <- file.path(system.file("extdata", package = "ollamar"), "image1.png")
image_encode_base64(image_path)
}
