% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last_response}
\alias{last_response}
\title{Get last response}
\usage{
last_response()
}
\value{
Last httr2 response or NULL.
}
\description{
Get and print the last response for debugging or when catching errors.
}
\examples{
test_connection(logical = TRUE)
last_response()
}
