% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{model_avail}
\alias{model_avail}
\title{Check if model is available locally}
\usage{
model_avail(model)
}
\arguments{
\item{model}{A character string of the model name such as "llama3".}
}
\value{
A logical value indicating if the model exists.
}
\description{
Check if model is available locally
}
\examples{
\dontshow{if (test_connection(logical = TRUE)) withAutoprint(\{ # examplesIf}
model_avail("codegemma:7b")
model_avail("abc")
model_avail("llama3")
\dontshow{\}) # examplesIf}
}
