% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_options.R
\docType{data}
\name{model_options}
\alias{model_options}
\title{Model options}
\format{
An object of class \code{list} of length 13.
}
\usage{
model_options
}
\description{
Model options
}
\keyword{datasets}
