% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\docType{data}
\name{package_config}
\alias{package_config}
\title{Package configuration}
\format{
An object of class \code{list} of length 3.
}
\usage{
package_config
}
\description{
Package configuration
}
\keyword{datasets}
