% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepend_message}
\alias{prepend_message}
\title{Prepend message to a list}
\usage{
prepend_message(content, role = "user", x = NULL)
}
\arguments{
\item{content}{The content of the message.}

\item{role}{The role of the message. Can be "user", "system", "assistant".}

\item{x}{A list of messages. Default is NULL.}
}
\value{
A list of messages with the new message prepended.
}
\description{
Prepends a message (add to beginning of a list) to a list of messages.
The role and content will be converted to a list and prepended to the input list.
}
\examples{
prepend_message("user", "Hello")
prepend_message("system", "Always respond nicely")
}
