% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_options.R
\name{search_options}
\alias{search_options}
\title{Search for options based on a query.}
\usage{
search_options(query)
}
\arguments{
\item{query}{A query (character) to search for in the options.}
}
\value{
Returns a list of matching options.
}
\description{
Search for options based on a query.
}
\examples{
search_options("learning rate")
search_options("tokens")
search_options("invalid query")
}
