% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_connection}
\alias{test_connection}
\title{Test connection to Ollama server}
\usage{
test_connection(url = "http://localhost:11434", logical = FALSE)
}
\arguments{
\item{url}{The URL of the Ollama server. Default is http://localhost:11434}

\item{logical}{Logical. If TRUE, returns a boolean value. Default is FALSE.}
}
\value{
Boolean value or httr2 response object, where status_code is either 200 (success) or 503 (error).
}
\description{
Tests whether the Ollama server is running or not.
}
\examples{
test_connection(logical = TRUE)
test_connection("http://localhost:11434") # default url
test_connection("http://127.0.0.1:11434")
}
