% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_message}
\alias{validate_message}
\title{Validate a message}
\usage{
validate_message(message)
}
\arguments{
\item{message}{A list with a single message of list class.}
}
\value{
TRUE if message is valid, otherwise an error is thrown.
}
\description{
Validate a message to ensure it has the required fields and the correct data types for the \code{chat()} function.
}
\examples{
validate_message(list(role = "user", content = "Hello"))
}
