% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_messages}
\alias{validate_messages}
\title{Validate a list of messages}
\usage{
validate_messages(messages)
}
\arguments{
\item{messages}{A list of messages, each of list class.}
}
\value{
TRUE if all messages are valid, otherwise warning messages are printed and FALSE is returned.
}
\description{
Validate a list of messages to ensure they have the required fields and the correct data types for the \code{chat()} function.
}
\examples{
validate_messages(create_messages(
    create_message("Be friendly", "system"),
    create_message("Hello")
))
}
