% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{ver}
\alias{ver}
\title{Retrieve Ollama version}
\usage{
ver(endpoint = "/api/version", host = NULL)
}
\arguments{
\item{endpoint}{The endpoint to list the running models. Default is "/api/version".}

\item{host}{The base URL to use. Default is NULL, which uses Ollama's default base URL.}
}
\value{
A character string of the Ollama version.
}
\description{
Retrieve Ollama version
}
\examples{
\dontshow{if (test_connection(logical = TRUE)) withAutoprint(\{ # examplesIf}
ver()
\dontshow{\}) # examplesIf}
}
\references{
\href{https://github.com/ollama/ollama/blob/main/docs/api.md#version}{API documentation}
}
