% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{create_request}
\alias{create_request}
\title{Create a httr2 request object.}
\usage{
create_request(endpoint)
}
\arguments{
\item{endpoint}{The endpoint to create the request}
}
\value{
A httr2 request object.
}
\description{
Creates a httr2 request object with the base URL, headers and endpoint. Used by other functions in the package and not intended to be used directly.
}
