% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{create_request}
\alias{create_request}
\title{Create a httr2 request object}
\usage{
create_request(endpoint, host = NULL)
}
\arguments{
\item{endpoint}{The endpoint to create the request}

\item{host}{The base URL to use. Default is NULL, which uses http://127.0.0.1:11434}
}
\value{
A httr2 request object.
}
\description{
Creates a httr2 request object with base URL, headers and endpoint. Used by other functions in the package and not intended to be used directly.
}
\examples{
create_request("/api/tags")
create_request("/api/chat")
create_request("/api/embeddings")
}
