% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{delete}
\alias{delete}
\title{Delete a model}
\usage{
delete(model, endpoint = "/api/delete")
}
\arguments{
\item{model}{A character string of the model name such as "llama3".}

\item{endpoint}{The endpoint to delete the model. Default is "/api/delete".}
}
\value{
A httr2 response object.
}
\description{
Delete a model from your local machine that you downlaoded using the pull() function. To see which models are available, use the list_models() function.
}
\examples{
\dontrun{
delete("llama3")
}
}
