% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{list_models}
\alias{list_models}
\title{Get available local models}
\usage{
list_models(
  output = c("resp", "jsonlist", "raw", "df"),
  endpoint = "/api/tags"
)
}
\arguments{
\item{output}{The output format. Default is "resp". Other options are "jsonlist", "raw", "df".}

\item{endpoint}{The endpoint to get the models. Default is "/api/tags".}
}
\value{
A httr2 response object, json list, raw or data frame.
}
\description{
Get available local models
}
\examples{
\dontrun{
list_models()
list_models("resp")
list_models("jsonlist")
list_models("raw")
list_models("df")
}
}
