% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{list_models}
\alias{list_models}
\title{Get available local models}
\usage{
list_models(
  output = c("df", "resp", "jsonlist", "raw"),
  endpoint = "/api/tags"
)
}
\arguments{
\item{output}{The output format. Default is "df". Other options are "resp", "jsonlist", "raw".}

\item{endpoint}{The endpoint to get the models. Default is "/api/tags".}
}
\value{
A httr2 response object, json list, raw or data frame. Default is "df".
}
\description{
Get available local models
}
\examples{
\dontshow{if (test_connection()$status_code == 200) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
list_models()  # returns dataframe/tibble by default
list_models("df")
list_models("resp")
list_models("jsonlist")
list_models("raw")
\dontshow{\}) # examplesIf}
}
