% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{ohelp}
\alias{ohelp}
\title{Chat with a model in real-time in R console}
\usage{
ohelp(model = "codegemma:7b", ...)
}
\arguments{
\item{model}{A character string of the model name such as "llama3". Defaults to "codegemma:7b" which is a decent coding model as of 2024-07-27.}

\item{...}{Additional options. No options are currently available at this time.}
}
\value{
Does not return anything. It prints the conversation in the console.
}
\description{
Chat with a model in real-time in R console
}
\examples{
\dontshow{if (test_connection()$status_code == 200) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ohelp(first_prompt = "quit")
# regular usage: ohelp()
\dontshow{\}) # examplesIf}
}
