% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{ps}
\alias{ps}
\title{List models that are currently loaded into memory}
\usage{
ps(
  output = c("df", "resp", "jsonlist", "raw", "text"),
  endpoint = "/api/ps",
  host = NULL
)
}
\arguments{
\item{output}{The output format. Default is "df". Supported formats are "df", "resp", "jsonlist", "raw", and "text".}

\item{endpoint}{The endpoint to list the running models. Default is "/api/ps".}

\item{host}{The base URL to use. Default is NULL, which uses Ollama's default base URL.}
}
\value{
A response in the format specified in the output parameter.
}
\description{
List models that are currently loaded into memory
}
\examples{
\dontshow{if (test_connection(logical = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ps("text")
\dontshow{\}) # examplesIf}
}
\references{
\href{https://github.com/ollama/ollama/blob/main/docs/api.md#list-running-models}{API documentation}
}
