% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{pull}
\alias{pull}
\title{Pull/download a model}
\usage{
pull(model, stream = TRUE, endpoint = "/api/pull")
}
\arguments{
\item{model}{A character string of the model name such as "llama3".}

\item{stream}{Enable response streaming. Default is TRUE.}

\item{endpoint}{The endpoint to pull the model. Default is "/api/pull".}
}
\value{
A httr2 response object.
}
\description{
See https://ollama.com/library for a list of available models. Use the list_models() function to get the list of models already downloaded/installed on your machine.
}
\examples{
\dontshow{if (test_connection()$status_code == 200) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pull("llama3")
\dontshow{\}) # examplesIf}
}
