% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resp_process}
\alias{resp_process}
\title{Process httr2 response object.}
\usage{
resp_process(resp, output = c("df", "jsonlist", "raw", "resp"))
}
\arguments{
\item{resp}{A httr2 response object.}

\item{output}{The output format. Default is "df". Other options are "jsonlist", "raw", "resp" (httr2 response object).}
}
\value{
A data frame, json list, raw or httr2 response object.
}
\description{
Process httr2 response object.
}
\examples{
resp <- list_models()
resp_process(resp)
resp_process(resp, "df")
resp_process(resp, "jsonlist")
resp_process(resp, "raw")
resp_process(resp, "resp")
}
