% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_connection}
\alias{test_connection}
\title{Test connection to Ollama server}
\usage{
test_connection(url = "http://localhost:11434")
}
\arguments{
\item{url}{The URL of the Ollama server. Default is http://localhost:11434}
}
\value{
A httr2 response object.
}
\description{
Test connection to Ollama server
}
\examples{
test_connection()
test_connection("http://localhost:11434") # default url
test_connection("http://127.0.0.1:11434")
}
